onlt use in interactive shells
git-svn-id: file:///usr/local/svn/jabs@563 f6a8b572-8bf8-43d5-8295-329fc01c5294

+++ bashrc
@@ -1,22 +1,25 @@
 #! /bin/sh
 
-. environ
-. python-environ
-test -f local/environ && . local/environ
-test -n "$PS1" || return
-. prompt green
-test -f local/prompt && . local/prompt
-. aliases
-test -f local/aliases && . local/aliases
-. functons
-test -f local/functons && . local/functons
-. bin/j.sh
-/bin/rm -rf tmp/*
-if [[ $USER != "builder" ]]
+if [[ $- == *i* ]] # http://www.gnu.org/software/bash/manual/html_node/Is-this-Shell-Interactive_003f.html#Is-this-Shell-Interactive_003f
 then
-	test -f todo.txt && cat todo.txt
-	echo
-	echo Welcome jab, to $HOSTNAME
-	echo
+	. environ
+	. python-environ
+	test -f local/environ && . local/environ
+	test -n "$PS1" || return
+	. prompt green
+	test -f local/prompt && . local/prompt
+	. aliases
+	test -f local/aliases && . local/aliases
+	. functons
+	test -f local/functons && . local/functons
+	. bin/j.sh
+	/bin/rm -rf tmp/*
+	if [[ $USER != "builder" ]]
+	then
+		test -f todo.txt && cat todo.txt
+		echo
+		echo Welcome jab, to $HOSTNAME
+		echo
+	fi
+	test -f ~/.oldpwd && . ~/.oldpwd
 fi
-test -f ~/.oldpwd && . ~/.oldpwd
